feat: enhance video streaming functionality with timestamped JPEG payloads, optimize frame processing, and update database schema for bigint support

+++ backend/app/db/models.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from uuid import UUID
 
-from sqlalchemy import DateTime, ForeignKey, Integer, Numeric, String
+from sqlalchemy import BigInteger, DateTime, ForeignKey, Integer, Numeric, String
 from sqlalchemy.dialects.postgresql import UUID as PgUUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
@@ -15,7 +15,7 @@ class StreamSession(UuidPrimaryKeyMixin, TimestampMixin, Base):
     source: Mapped[str] = mapped_column(String(64), nullable=False, default="browser")
     started_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False)
     ended_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True)
-    frame_count: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
+    frame_count: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)
 
     observations: Mapped[list["RoiObservation"]] = relationship(
         back_populates="session",
@@ -32,8 +32,8 @@ class RoiObservation(UuidPrimaryKeyMixin, TimestampMixin, Base):
         nullable=False,
         index=True,
     )
-    frame_number: Mapped[int] = mapped_column(Integer, nullable=False, index=True)
-    timestamp_ms: Mapped[int] = mapped_column(Integer, nullable=False)
+    frame_number: Mapped[int] = mapped_column(BigInteger, nullable=False, index=True)
+    timestamp_ms: Mapped[int] = mapped_column(BigInteger, nullable=False)
     x: Mapped[int] = mapped_column(Integer, nullable=False)
     y: Mapped[int] = mapped_column(Integer, nullable=False)
     width: Mapped[int] = mapped_column(Integer, nullable=False)
